+++ MESManager/MESManager.Infrastructure/DependencyInjection.cs
@@ -42,6 +42,7 @@ public static class DependencyInjection
         services.AddScoped<IOperatoreAppService, OperatoreAppService>();
         services.AddScoped<ICalendarioLavoroAppService, CalendarioLavoroAppService>();
         services.AddScoped<IImpostazioniGanttAppService, ImpostazioniGanttAppService>();
+        services.AddScoped<IFestiviAppService, FestiviAppService>();
         services.AddScoped<IUtenteAppService, UtenteAppService>();
         services.AddScoped<IPreferenzeUtenteService, PreferenzeUtenteService>();
         services.AddScoped<CurrentUserService>();